# docs/swagger.yaml
basePath: /v1
definitions:
...
  gosnmp.SnmpPDU:
    properties:
      name:
        type: string
      type:
        type: Asn1BER
      value:
        type: object
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: lyonsdpy@163.com
    name: 代澎源
    url: https://github.com/lyonsdpy/phoenix
  description: |-
    可用于监控采集,网管信息获取的统一入口.
    后续计划支持prometheus，grpc，etcd.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: phoenix采集agent
  version: 1.0.0
paths:
...
  /snmp-bulkget/sync:
    post:
      consumes:
      - application/json
      description: SNMP的交互
      parameters:
      - default: 127.0.0.1:161
        description: snmp请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 团体字
        in: header
        name: community
        required: true
        type: string
      - description: OID列表
        in: body
        name: oid_list
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{1:"integer",2:"bitstring"...}'
          schema:
            items:
              $ref: '#/definitions/gosnmp.SnmpPDU'
            type: array
      tags:
      - SNMP
  /snmp-typelist:
    post:
      consumes:
      - application/json
      description: SNMP的交互
      parameters:
      - default: 127.0.0.1:161
        description: snmp请求目标
        in: header
        name: target
        required: true
        type: string
      - description: 团体字
        in: header
        name: community
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{1:"integer",2:"bitstring"...}'
          schema:
            type: string
      tags:
      - SNMP
swagger: "2.0"